# Frontend/RepoGeist/openapi.json
{
    "openapi":  "3.1.1",
    "info":  {
                 "title":  "RepoGeist | v1",
                 "version":  "1.0.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:5221/"
                    }
                ],
    "paths":  {
...
                  "/api/repos/{owner}/{name}/commit-activity":  {
                                                                    "get":  {
                                                                                "tags":  [
                                                                                             "Health"
                                                                                         ],
                                                                                "parameters":  [
                                                                                                   {
                                                                                                       "name":  "owner",
                                                                                                       "in":  "path",
                                                                                                       "required":  true,
                                                                                                       "schema":  {
                                                                                                                      "type":  "string"
                                                                                                                  }
                                                                                                   },
                                                                                                   {
                                                                                                       "name":  "name",
                                                                                                       "in":  "path",
                                                                                                       "required":  true,
                                                                                                       "schema":  {
                                                                                                                      "type":  "string"
                                                                                                                  }
                                                                                                   },
                                                                                                   {
                                                                                                       "name":  "weeks",
                                                                                                       "in":  "query",
                                                                                                       "schema":  {
                                                                                                                      "pattern":  "^-?(?:0|[1-9]\\d*)$",
                                                                                                                      "type":  [
                                                                                                                                   "integer",
                                                                                                                                   "string"
                                                                                                                               ],
                                                                                                                      "format":  "int32",
                                                                                                                      "default":  52
                                                                                                                  }
                                                                                                   }
                                                                                               ],
                                                                                "responses":  {
                                                                                                  "200":  {
                                                                                                              "description":  "OK",
                                                                                                              "content":  {
                                                                                                                              "text/plain":  {
                                                                                                                                                 "schema":  {
                                                                                                                                                                "type":  "array",
                                                                                                                                                                "items":  {
                                                                                                                                                                              "$ref":  "#/components/schemas/CommitActivityDto"
                                                                                                                                                                          }
                                                                                                                                                            }
                                                                                                                                             },
                                                                                                                              "application/json":  {
                                                                                                                                                       "schema":  {
                                                                                                                                                                      "type":  "array",
                                                                                                                                                                      "items":  {
                                                                                                                                                                                    "$ref":  "#/components/schemas/CommitActivityDto"
                                                                                                                                                                                }
                                                                                                                                                                  }
                                                                                                                                                   },
                                                                                                                              "text/json":  {
                                                                                                                                                "schema":  {
                                                                                                                                                               "type":  "array",
                                                                                                                                                               "items":  {
                                                                                                                                                                             "$ref":  "#/components/schemas/CommitActivityDto"
                                                                                                                                                                         }
                                                                                                                                                           }
                                                                                                                                            }
                                                                                                                          }
                                                                                                          }
                                                                                              }
                                                                            }
                                                                },
...
                  "/api/repos/{owner}/{name}":  {
                                                    "get":  {
                                                                "tags":  [
                                                                             "Repos"
                                                                         ],
                                                                "parameters":  [
                                                                                   {
                                                                                       "name":  "owner",
                                                                                       "in":  "path",
                                                                                       "required":  true,
                                                                                       "schema":  {
                                                                                                      "type":  "string"
                                                                                                  }
                                                                                   },
                                                                                   {
                                                                                       "name":  "name",
                                                                                       "in":  "path",
                                                                                       "required":  true,
                                                                                       "schema":  {
                                                                                                      "type":  "string"
                                                                                                  }
                                                                                   }
                                                                               ],
                                                                "responses":  {
                                                                                  "200":  {
                                                                                              "description":  "OK",
                                                                                              "content":  {
                                                                                                              "text/plain":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "$ref":  "#/components/schemas/RepoDetailDto"
                                                                                                                                            }
                                                                                                                             },
                                                                                                              "application/json":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/RepoDetailDto"
                                                                                                                                                  }
                                                                                                                                   },
                                                                                                              "text/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "$ref":  "#/components/schemas/RepoDetailDto"
                                                                                                                                           }
                                                                                                                            }
                                                                                                          }
                                                                                          }
                                                                              }
                                                            }
                                                },
...
              },
    "components":  {
                       "schemas":  {
                                       "AnalysisStatus":  {
                                                              "type":  "integer"
                                                          },
...
                                       "CommitActivityDto":  {
                                                                 "type":  "object",
                                                                 "properties":  {
                                                                                    "weekStart":  {
                                                                                                      "type":  "string",
                                                                                                      "format":  "date-time"
                                                                                                  },
                                                                                    "commitCount":  {
                                                                                                        "pattern":  "^-?(?:0|[1-9]\\d*)$",
                                                                                                        "type":  [
                                                                                                                     "integer",
                                                                                                                     "string"
                                                                                                                 ],
                                                                                                        "format":  "int32"
                                                                                                    }
                                                                                }
                                                             },
                                       "ContributorDto":  {
                                                              "type":  "object",
                                                              "properties":  {
                                                                                 "githubUsername":  {
                                                                                                        "type":  "string"
                                                                                                    },
                                                                                 "avatarUrl":  {
                                                                                                   "type":  "string"
                                                                                               },
                                                                                 "commitCount":  {
                                                                                                     "pattern":  "^-?(?:0|[1-9]\\d*)$",
                                                                                                     "type":  [
                                                                                                                  "integer",
                                                                                                                  "string"
                                                                                                              ],
                                                                                                     "format":  "int32"
                                                                                                 }
                                                                             }
                                                          },
...
                                       "HealthStatus":  {
                                                            "type":  "integer"
                                                        },
...
                                       "LanguageStatDto":  {
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "language":  {
                                                                                                   "type":  "string"
                                                                                               },
                                                                                  "byteCount":  {
                                                                                                    "pattern":  "^-?(?:0|[1-9]\\d*)$",
                                                                                                    "type":  [
                                                                                                                 "integer",
                                                                                                                 "string"
                                                                                                             ],
                                                                                                    "format":  "int64"
                                                                                                }
                                                                              }
                                                           },
...
                                       "RepoDetailDto":  {
                                                             "type":  "object",
                                                             "properties":  {
                                                                                "id":  {
                                                                                           "type":  "string",
                                                                                           "format":  "uuid"
                                                                                       },
                                                                                "owner":  {
                                                                                              "type":  "string"
                                                                                          },
                                                                                "name":  {
                                                                                             "type":  "string"
                                                                                         },
                                                                                "description":  {
                                                                                                    "type":  [
                                                                                                                 "null",
                                                                                                                 "string"
                                                                                                             ]
                                                                                                },
                                                                                "topics":  {
                                                                                               "type":  "array",
                                                                                               "items":  {
                                                                                                             "type":  "string"
                                                                                                         }
                                                                                           },
                                                                                "readmeHtml":  {
                                                                                                   "type":  [
                                                                                                                "null",
                                                                                                                "string"
                                                                                                            ]
                                                                                               },
                                                                                "languages":  {
                                                                                                  "type":  "array",
                                                                                                  "items":  {
                                                                                                                "$ref":  "#/components/schemas/LanguageStatDto"
                                                                                                            }
                                                                                              },
                                                                                "contributors":  {
                                                                                                     "type":  "array",
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ContributorDto"
                                                                                                               }
                                                                                                 },
                                                                                "commitActivity":  {
                                                                                                       "type":  "array",
                                                                                                       "items":  {
                                                                                                                     "$ref":  "#/components/schemas/CommitActivityDto"
                                                                                                                 }
                                                                                                   },
                                                                                "entryPointPath":  {
                                                                                                       "type":  [
                                                                                                                    "null",
                                                                                                                    "string"
                                                                                                                ]
                                                                                                   },
                                                                                "liveDemoUrl":  {
                                                                                                    "type":  [
                                                                                                                 "null",
                                                                                                                 "string"
                                                                                                             ]
                                                                                                },
                                                                                "healthStatus":  {
                                                                                                     "$ref":  "#/components/schemas/HealthStatus"
                                                                                                 },
                                                                                "stars":  {
                                                                                              "pattern":  "^-?(?:0|[1-9]\\d*)$",
                                                                                              "type":  [
                                                                                                           "integer",
                                                                                                           "string"
                                                                                                       ],
                                                                                              "format":  "int32"
                                                                                          },
                                                                                "forks":  {
                                                                                              "pattern":  "^-?(?:0|[1-9]\\d*)$",
                                                                                              "type":  [
                                                                                                           "integer",
                                                                                                           "string"
                                                                                                       ],
                                                                                              "format":  "int32"
                                                                                          },
                                                                                "analysisStatus":  {
                                                                                                       "$ref":  "#/components/schemas/AnalysisStatus"
                                                                                                   }
                                                                            }
                                                         }
                                   }
                   },
...
}